Temp files are now more controlled
feat: Added an option to remove or not the temp files at the end of the program.
feat: If the temp output folder is not empty at the start of the application, the app empties it.

+++ src/utilities/DataFusion.py
@@ -35,11 +35,12 @@ class Util:
     def sortByIdOfProduct(list):
         list.sort(key=lambda x : x["IDProduct"])
         
-    def cleanTempFiles():
+    def cleanTempFiles(forceClean:bool = False):
         '''Removes the temporary files created by the blackboxes.'''
-        pathFile =  OC.getTempData()
-        shutil.rmtree(pathFile,ignore_errors=True)
-        os.mkdir(pathFile)
+        if OC.getKeepTempFiles()==False or forceClean==True:
+            pathFile =  OC.getTempData()
+            shutil.rmtree(pathFile,ignore_errors=True)
+            os.mkdir(pathFile)
     
     def registerProductAsDone(product:str):
         '''Registers the product in a file. Products found in this file will not be dealt with again.'''
